Support passing variables in test
